abstrqcted out accounts div inti it's own component
